Move Symbol Table into TuffScript package
