+++ packages/tuffscript/src/symbolTable/index.ts
